twitter_api: clarify tweet encoding stuff
There were a number of little confusing things that took a long time to
explain during the Twitter module run-through that are made better here.

First, this commit eliminates the need to have different code
on Windows versus other platforms.

Second, it changes the name of print_tweet() to look less like 'print
tweet' (using the print statement instead of calling a function).

Third, it eliminates the usage of 'print' within a called function,
which made it harder to modify what was printed later.